requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
flask-login>=0.6.3
cryptography>=42.0.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0
openai>=1.0.0
celery[redis]>=5.3.0
redis>=5.0.0
python-docx>=1.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0
markdown>=3.5.0
PyMySQL>=1.1.0